requirements.txt
requests>=2.31.0
onnxruntime>=1.17.0
Pillow>=10.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
huggingface_hub>=0.20.0
pytest>=7.0.0
customtkinter>=5.2.0
